requirements.txt
certifi==2022.6.15
python-dotenv==0.20.0
requests==2.28.0
urllib3==1.26.9
